app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
coverage==7.7.0
exceptiongroup==1.2.2
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
kafka-python==2.1.2
packaging==24.2
pip==23.0.1
pluggy==1.5.0
prometheus_client==0.21.1
prometheus-fastapi-instrumentator==7.1.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings==2.8.1
pytest==8.3.5
pytest-mock==3.14.0
python-dotenv==1.0.1
setuptools==65.5.1
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
wheel==0.45.1
